+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
-<!--
-## Next_Ver - 2022/xx/xx
--->
+## 20240330
+
+- Update version, screenshots
+- Update UI - add buffer representation
 
 ## 20230828